+++ R/proxy.R
@@ -29,6 +29,31 @@
 #'   therefore larger.
 #' @param digits determines rounding of small values towards zero. Use primarily
 #'   to correct rounding errors in C++. See \link{zapsmall}.
+#' @details
+#' Available methods for similarity:
+#' \itemize{
+#'   \item `cosine`: cosine similarity
+#'   \item `correlation`: Pearson's correlation
+#'   \item `jaccard`: Jaccard coefficient
+#'   \item `ejaccard`: the real value version of `jaccard`
+#'   \item `dice`: Dice coefficient
+#'   \item `edice`: the real value version of `dice`
+#'   \item `faith`: Faith similarity
+#'   \item `simple matching`: the percentage of common elements
+#' }
+#' Available methods for distance:
+#' \itemize{
+#'   \item `euclidean`: Euclidean distance
+#'   \item `chisquared`: chi-squared distance
+#'   \item `kullback`: Kullback–Leibler divergence
+#'   \item `jeffreys`: Jeffreys divergence
+#'   \item `jensen`: Jensen–Shannon divergence
+#'   \item `manhattan`: Manhattan distance
+#'   \item `maximum`: the largest difference between values
+#'   \item `canberra`: Canberra distance
+#'   \item `minkowski`: Minkowski distance
+#'   \item `hamming`: Hamming distance
+#' }
 #' @import methods Matrix
 #' @importFrom RcppParallel RcppParallelLibs
 #' @seealso zapsmall
@@ -38,7 +63,7 @@
 #' simil(mt, method = "cosine")[1:5, 1:5]
 simil <- function(x, y = NULL, margin = 1,
                   method = c("cosine", "correlation", "jaccard", "ejaccard",
-                             "dice", "edice", "hamann", "simple matching", "faith"),
+                             "dice", "edice", "hamann", "faith", "simple matching"),
                   min_simil = NULL, rank = NULL, drop0 = FALSE, diag = FALSE,
                   use_nan = FALSE, digits = 14) {
 
@@ -52,12 +77,13 @@ simil <- function(x, y = NULL, margin = 1,
 #' @rdname simil
 #' @param smooth adds a  fixed value to all the cells to avoid division by zero.
 #'   Only used when `method` is "chisquared", "kullback" or "jeffreys".
+
 #' @export
 #' @examples
 #' mt <- Matrix::rsparsematrix(100, 100, 0.01)
 #' dist(mt, method = "euclidean")[1:5, 1:5]
 dist <- function(x, y = NULL, margin = 1,
-                 method = c("euclidean", "chisquared", "kullback", "jeffreys",
+                 method = c("euclidean", "chisquared", "kullback", "jeffreys", "jensen",
                             "manhattan", "maximum", "canberra", "minkowski", "hamming"),
                  p = 2, smooth = 0, drop0 = FALSE, diag = FALSE, use_nan = FALSE, digits = 14) {
 
@@ -71,7 +97,7 @@ dist <- function(x, y = NULL, margin = 1,
 proxy <- function(x, y = NULL, margin = 1,
                   method = c("cosine", "correlation", "jaccard", "ejaccard",
                              "dice", "edice", "hamann", "simple matching", "faith",
-                             "euclidean", "chisquared", "kullback", "jeffreys",
+                             "euclidean", "chisquared", "kullback", "jeffreys", "jensen",
                              "manhattan", "maximum", "canberra", "minkowski", "hamming"),
                   p = 2, smooth = 0, min_proxy = NULL, rank = NULL, drop0 = FALSE,
                   diag = FALSE, use_nan = FALSE, digits = 14) {
@@ -108,7 +134,7 @@ proxy <- function(x, y = NULL, margin = 1,
         if (method == "correlation") {
             if (any(colSds(x) == 0) || any(colSds(y) == 0))
                 warning("x or y has vectors with zero standard deviation; consider setting use_nan = TRUE", call. = FALSE)
-        } else if (method %in% c("cosine", "kullback", "chisquared", "jeffreys")) {
+        } else if (method %in% c("cosine", "kullback", "chisquared", "jeffreys", "jensen")) {
             if (any(colZeros(x) == nrow(x)) || any(colZeros(y) == nrow(y)))
                 warning("x or y has vectors with all zero; consider setting use_nan = TRUE", call. = FALSE)
         }
@@ -159,7 +185,7 @@ proxy <- function(x, y = NULL, margin = 1,
                                      "hamann", "simple matching", "faith",
                                      "euclidean", "chisquared", "kullback", "manhattan",
                                      "maximum", "canberra", "minkowski", "hamming",
-                                     "jeffreys")),
+                                     "jeffreys", "jensen")),
             rank = rank,
             limit = min_proxy,
             weight = weight,